requirements.txt
black==23.1.0
celery>=4.4.0
Coverage==6.3.2
Django==3.2.9
django-celery-beat>=2.2.0
django-cors-headers>=3.7.0
django-extensions>=3.1.3
django-filter==21.1
djangorestframework>=3.12.2
djangorestframework-simplejwt==4.8.0
drf-yasg>=1.20.0
factory-boy>=2.9.0
flower==1.0.0
mock>=4.0.3
pyfcm>=1.5.4
pytest>=6.2.5
pytest-django>=4.1.0
pytest-cov==3.0.0
psycopg2>=2.8.5